requirements.txt
fastapi
uvicorn
lancedb
pandas
openai
python-multipart
requests
python-dotenv
aiofiles
